Split saltwater into crates
