+++ Cargo.toml
@@ -21,27 +21,28 @@ default-run = "swcc"
 documentation = "https://docs.rs/saltwater"
 
 [dependencies]
-lazy_static = "1"
-ansi_term = { version = "0.12", optional = true }
-cranelift = { version = "0.66", optional = true }
-cranelift-module = { version = "0.66", optional = true }
-cranelift-object = { version = "0.66", optional = true }
-cranelift-simplejit = { version = "0.66", optional = true }
-hexponent = "0.3"
-thiserror = "^1.0.10"
-target-lexicon = "0.10"
-tempfile = { version = "3", optional = true }
-pico-args = { version = "0.3.4", optional = true, features = ["short-space-opt"] }
-lasso = "0.2"
-codespan = { version = ">=0.9.5", default-features = false }
+#lazy_static = "1"
+ansi_term = "0.12"
+arcstr = "0.2"
+#cranelift = { version = "0.66", optional = true }
+#cranelift-module = { version = "0.66", optional = true }
+#cranelift-object = { version = "0.66", optional = true }
+#cranelift-simplejit = { version = "0.66", optional = true }
+#hexponent = "0.3"
+#thiserror = "^1.0.10"
+#target-lexicon = "0.10"
+tempfile = "3"
+pico-args = { version = "0.3.4", features = ["short-space-opt"] }
+#lasso = "0.2"
+#codespan = { version = ">=0.9.5", default-features = false }
 color-backtrace = { version = "0.4", default-features = false, optional = true }
-counter = "0.4"
-atty = { version = "0.2", default-features = false, optional = true }
+#counter = "0.4"
+atty = { version = "0.2", default-features = false }
 git-testament = { version = "0.1", optional = true }
 rand = { version = "0.7", optional = true }
 rodio = { version = "0.11.0", optional = true }
-arcstr = "0.2.2"
-time = "0.2"
+saltwater-parser = { path = "saltwater-parser" }
+saltwater-codegen = { path = "saltwater-codegen" }
 
 [dev-dependencies]
 env_logger = { version = "0.7", default-features = false }
@@ -52,11 +53,8 @@ proptest = "^0.9.6"
 proptest-derive = "0.1"
 
 [features]
-default = ["cc", "codegen", "color-backtrace"]
-# The `swcc` binary
-cc = ["ansi_term", "tempfile", "pico-args", "codegen", "atty"]
-codegen = ["cranelift", "cranelift-module", "cranelift-object"]
-jit = ["codegen", "cranelift-simplejit"]
+default = ["color-backtrace"]
+#codegen = ["cranelift", "cranelift-module", "cranelift-object"]
 salty = ["rand", "rodio"]
 # for internal use
 _test_headers = []
@@ -64,7 +62,6 @@ _test_headers = []
 [[bin]]
 name = "swcc"
 path = "src/main.rs"
-required-features = ["cc"]
 
 [[bench]]
 name = "examples"
