+++ saltwater-codegen/lib.rs
@@ -14,42 +14,33 @@ mod expr;
 mod static_init;
 mod stmt;
 
-use std::collections::{HashMap, VecDeque};
+use std::collections::HashMap;
 use std::convert::TryFrom;
+use std::path::Path;
 
-use crate::arch::{CHAR_BIT, PTR_SIZE, SIZE_T, TARGET};
-use crate::data::lex::ComparisonToken;
+use saltwater_parser::{Program, Opt};
+use saltwater_parser::arch::TARGET;
 use cranelift::codegen::{
     self,
     ir::{
-        condcodes::{FloatCC, IntCC},
         entities::StackSlot,
         function::Function,
         stackslot::{StackSlotData, StackSlotKind},
-        types::{self, Type as IrType},
-        AbiParam, ArgumentPurpose, ExternalName, InstBuilder, MemFlags, Signature,
+        ExternalName, InstBuilder, MemFlags,
     },
-    isa::{CallConv, TargetIsa},
+    isa::TargetIsa,
     settings::{self, Configurable, Flags},
 };
 use cranelift::frontend::Switch;
 use cranelift::prelude::{Block, FunctionBuilder, FunctionBuilderContext};
 use cranelift_module::{self, Backend, DataId, FuncId, Linkage, Module};
 use cranelift_object::{ObjectBackend, ObjectBuilder};
-use lazy_static::lazy_static;
 
-use crate::data::{
+use saltwater_parser::data::{
     hir::{Declaration, Initializer, Stmt, Symbol},
     types::FunctionType,
     StorageClass, *,
 };
-
-// TODO: make this const when const_if_match is stabilized
-// TODO: see https://github.com/rust-lang/rust/issues/49146
-lazy_static! {
-    /// The calling convention for the current target.
-    pub(crate) static ref CALLING_CONVENTION: CallConv = CallConv::triple_default(&TARGET);
-}
 
 pub(crate) fn get_isa(jit: bool) -> Box<dyn TargetIsa + 'static> {
     let mut flags_builder = cranelift::codegen::settings::builder();
@@ -105,48 +96,15 @@ struct Compiler<T: Backend> {
     error_handler: ErrorHandler,
 }
 
+/*
 /// Compile a program from a high level IR to a Cranelift Module
 pub(crate) fn compile<B: Backend>(
     module: Module<B>,
     program: Vec<Locatable<Declaration>>,
     debug: bool,
 ) -> (Result<Module<B>, CompileError>, VecDeque<CompileWarning>) {
-    // really we'd like to have all errors but that requires a refactor
-    let mut err = None;
-    let mut compiler = Compiler::new(module, debug);
-    for decl in program {
-        let meta = decl.data.symbol.get();
-        if let StorageClass::Typedef = meta.storage_class {
-            continue;
-        }
-        let current = match &meta.ctype {
-            Type::Function(func_type) => match decl.data.init {
-                Some(Initializer::FunctionBody(stmts)) => {
-                    compiler.compile_func(decl.data.symbol, &func_type, stmts, decl.location)
-                }
-                None => compiler.declare_func(decl.data.symbol, false).map(|_| ()),
-                _ => unreachable!("functions can only be initialized by a FunctionBody"),
-            },
-            Type::Void | Type::Error => unreachable!("parser let an incomplete type through"),
-            _ => {
-                if let Some(Initializer::FunctionBody(_)) = &decl.data.init {
-                    unreachable!("only functions should have a function body")
-                }
-                compiler.store_static(decl.data.symbol, decl.data.init, decl.location)
-            }
-        };
-        if let Err(e) = current {
-            err = Some(e);
-            break;
-        }
-    }
-    let warns = compiler.error_handler.warnings;
-    if let Some(err) = err {
-        (Err(err), warns)
-    } else {
-        (Ok(compiler.module), warns)
-    }
 }
+*/
 
 impl<B: Backend> Compiler<B> {
     fn new(module: Module<B>, debug: bool) -> Compiler<B> {
@@ -173,7 +131,7 @@ impl<B: Backend> Compiler<B> {
     // 3. should always declare `id` as export or local.
     // 2. and 4. should be a no-op.
     fn declare_func(&mut self, symbol: Symbol, is_definition: bool) -> CompileResult<FuncId> {
-        use crate::get_str;
+        use saltwater_parser::get_str;
         if !is_definition {
             // case 2 and 4
             if let Some(Id::Function(func_id)) = self.declarations.get(&symbol) {
@@ -395,133 +353,99 @@ impl<B: Backend> Compiler<B> {
     }
 }
 
-impl FunctionType {
-    fn has_params(&self) -> bool {
-        !(self.params.len() == 1 && self.params[0].get().ctype == Type::Void)
-    }
+pub type Product = <cranelift_object::ObjectBackend as Backend>::Product;
 
-    /// Generate the IR function signature for `self`
-    pub fn signature(&self, isa: &dyn TargetIsa) -> Signature {
-        let mut params = if self.params.len() == 1 && self.params[0].get().ctype == Type::Void {
-            // no arguments
-            Vec::new()
-        } else {
-            self.params
-                .iter()
-                .map(|param| AbiParam::new(param.get().ctype.as_ir_type()))
-                .collect()
-        };
-        if self.varargs {
-            let al = isa
-                .register_info()
-                .parse_regunit("rax")
-                .expect("x86 should have an rax register");
-            params.push(AbiParam::special_reg(
-                types::I8,
-                ArgumentPurpose::Normal,
-                al,
-            ));
-        }
-        let return_type = if !self.should_return() {
-            vec![]
-        } else {
-            vec![AbiParam::new(self.return_type.as_ir_type())]
-        };
-        Signature {
-            call_conv: *CALLING_CONVENTION,
-            params,
-            returns: return_type,
-        }
-    }
-}
+/// Compile and return the declarations and warnings.
+pub fn compile<B: Backend>(module: Module<B>, buf: &str, opt: Opt) -> Program<Module<B>> {
+    use saltwater_parser::{vec_deque, check_semantics};
 
-impl ComparisonToken {
-    pub fn to_int_compare(self, signed: bool) -> IntCC {
-        use ComparisonToken::*;
-        match (self, signed) {
-            (Less, true) => IntCC::SignedLessThan,
-            (Less, false) => IntCC::UnsignedLessThan,
-            (LessEqual, true) => IntCC::SignedLessThanOrEqual,
-            (LessEqual, false) => IntCC::UnsignedLessThanOrEqual,
-            (Greater, true) => IntCC::SignedGreaterThan,
-            (Greater, false) => IntCC::UnsignedGreaterThan,
-            (GreaterEqual, true) => IntCC::SignedGreaterThanOrEqual,
-            (GreaterEqual, false) => IntCC::UnsignedGreaterThanOrEqual,
-            (EqualEqual, _) => IntCC::Equal,
-            (NotEqual, _) => IntCC::NotEqual,
-        }
-    }
-    pub fn to_float_compare(self) -> FloatCC {
-        use ComparisonToken::*;
-        match self {
-            Less => FloatCC::LessThan,
-            LessEqual => FloatCC::LessThanOrEqual,
-            Greater => FloatCC::GreaterThan,
-            GreaterEqual => FloatCC::GreaterThanOrEqual,
-            EqualEqual => FloatCC::Equal,
-            NotEqual => FloatCC::NotEqual,
-        }
-    }
-}
-
-use std::convert::TryInto;
-impl Type {
-    /// Return an IR integer type large enough to contain a pointer.
-    pub fn ptr_type() -> IrType {
-        IrType::int(CHAR_BIT * PTR_SIZE).expect("pointer size should be valid")
-    }
-    /// Return an IR type which can represent this C type
-    pub fn as_ir_type(&self) -> IrType {
-        use Type::*;
-
-        match self {
-            // Integers
-            Bool => types::B1,
-            Char(_) | Short(_) | Int(_) | Long(_) | Pointer(_, _) | Enum(_, _) => {
-                let int_size = SIZE_T::from(CHAR_BIT)
-                    * self
-                        .sizeof()
-                        .expect("integers should always have a valid size");
-                IrType::int(int_size.try_into().unwrap_or_else(|_| {
-                    panic!(
-                        "integers should never have a size larger than {}",
-                        i16::max_value()
-                    )
-                }))
-                .unwrap_or_else(|| panic!("unsupported size for IR: {}", int_size))
+    let debug_asm = opt.debug_asm;
+    let mut program = check_semantics(buf, opt);
+    let hir = match program.result {
+        Ok(hir) => hir,
+        Err(err) => {
+            return Program {
+                result: Err(err),
+                warnings: program.warnings,
+                files: program.files,
             }
-
-            // Floats
-            // TODO: this is hard-coded for x64
-            Float => types::F32,
-            Double => types::F64,
-
-            // Aggregates
-            // arrays and functions decay to pointers
-            Function(_) | Array(_, _) => IrType::int(PTR_SIZE * CHAR_BIT)
-                .unwrap_or_else(|| panic!("unsupported size of IR: {}", PTR_SIZE)),
-            // void cannot be loaded or stored
-            _ => types::INVALID,
+        }
+    };
+    // really we'd like to have all errors but that requires a refactor
+    let mut err = None;
+    let mut compiler = Compiler::new(module, debug_asm);
+    for decl in hir {
+        let meta = decl.data.symbol.get();
+        if let StorageClass::Typedef = meta.storage_class {
+            continue;
+        }
+        let current = match &meta.ctype {
+            Type::Function(func_type) => match decl.data.init {
+                Some(Initializer::FunctionBody(stmts)) => {
+                    compiler.compile_func(decl.data.symbol, &func_type, stmts, decl.location)
+                }
+                None => compiler.declare_func(decl.data.symbol, false).map(|_| ()),
+                _ => unreachable!("functions can only be initialized by a FunctionBody"),
+            },
+            Type::Void | Type::Error => unreachable!("parser let an incomplete type through"),
+            _ => {
+                if let Some(Initializer::FunctionBody(_)) = &decl.data.init {
+                    unreachable!("only functions should have a function body")
+                }
+                compiler.store_static(decl.data.symbol, decl.data.init, decl.location)
+            }
+        };
+        if let Err(e) = current {
+            err = Some(e);
+            break;
         }
     }
-    fn member_offset(&self, member: InternedStr) -> Result<u64, ()> {
-        match self {
-            Type::Struct(stype) => Ok(stype.offset(member)),
-            Type::Union(_) => Ok(0),
-            _ => Err(()),
-        }
+    let warns = compiler.error_handler.warnings;
+    let (result, ir_warnings) = if let Some(err) = err {
+        (Err(err), warns)
+    } else {
+        (Ok(compiler.module), warns)
+    };
+    program.warnings.extend(ir_warnings);
+    Program {
+        result: result.map_err(|errs| vec_deque![errs]),
+        warnings: program.warnings,
+        files: program.files,
     }
 }
 
-impl CompileError {
-    fn semantic(err: Locatable<String>) -> Self {
-        Self::from(err)
-    }
+pub fn assemble(product: Product, output: &Path) -> Result<(), saltwater_parser::Error> {
+    use std::io::{self, Write};
+    use std::fs::File;
+
+    let bytes = product.emit().map_err(saltwater_parser::Error::Platform)?;
+    File::create(output)?
+        .write_all(&bytes)
+        .map_err(io::Error::into)
 }
 
-impl FunctionType {
-    fn should_return(&self) -> bool {
-        *self.return_type != Type::Void
+pub fn link(obj_file: &Path, output: &Path) -> Result<(), std::io::Error> {
+    use std::io::{Error, ErrorKind};
+    use std::process::Command;
+
+    // link the .o file using host linker
+    let status = Command::new("cc")
+        .args(&[&obj_file, Path::new("-o"), output])
+        .status()
+        .map_err(|err| {
+            if err.kind() == ErrorKind::NotFound {
+                Error::new(
+                    ErrorKind::NotFound,
+                    "could not find host cc (for linking). Is it on your PATH?",
+                )
+            } else {
+                err
+            }
+        })?;
+    if !status.success() {
+        Err(Error::new(ErrorKind::Other, "linking program failed"))
+    } else {
+        Ok(())
     }
 }
