+++ saltwater-parser/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "saltwater-parser"
+version = "0.1.0"
+authors = ["Joshua Nelson <jyn514@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "lib.rs"
+
+[dependencies]
+lazy_static = "1"
+arcstr = "0.2.2"
+ansi_term = { version = "0.12", optional = true }
+hexponent = "0.3"
+thiserror = "^1.0.10"
+target-lexicon = "0.10"
+tempfile = { version = "3", optional = true }
+pico-args = { version = "0.3.4", optional = true, features = ["short-space-opt"] }
+lasso = "0.2"
+codespan = { version = ">=0.9.5", default-features = false }
+color-backtrace = { version = "0.4", default-features = false, optional = true }
+counter = "0.4"
+atty = { version = "0.2", default-features = false, optional = true }
+git-testament = { version = "0.1", optional = true }
+rand = { version = "0.7", optional = true }
+rodio = { version = "0.11.0", optional = true }
+shared_str = "0.1.1"
+time = "0.2"
+cranelift-codegen = { version = "0.66", optional = true }
+cranelift-object = { version = "0.66", optional = true }
+
+[features]
+codegen = ["cranelift-codegen", "cranelift-object"]
